#161 Move code generation methods from contributors to separate classes
